% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pgram.R
\name{calc_constants}
\alias{calc_constants}
\title{Produce required constants for periodogram calculation}
\usage{
calc_constants(interval, wear_length = 14)
}
\arguments{
\item{interval, }{time (in minutes) between CGM observations}

\item{wear_length, }{number of days users are expected to wear the device, set
by default to 14}
}
\value{
constant_list, named list of the required constants
}
\description{
Produce required constants for periodogram calculation
}
\details{
Calculates expected number of observations per-hour, per-wear, and
per-day
}
