% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCI.R
\name{calculate_GCI}
\alias{calculate_GCI}
\title{Calculates Glucose Color Index (GCI)}
\usage{
calculate_GCI(x, weights, means, stddevs, ...)
}
\arguments{
\item{x, }{dataframe with raw CGM time series or object of class "6DF_summary"
containing 6-DF representation of log-periodogram}

\item{weights, }{numeric vector of weights used to form the GCI linear combination}

\item{means, }{numeric vector of mean values for the 6 degree of freedom
approximation, subtracted away before calculation of GCI}

\item{stddevs, }{numeric vector of standard deviations for each of the values
within the 6 degree of freedom approximation, used for scaling before GCI calculation}

\item{..., }{additional arguments/settings required}
}
\value{
gci, numeric containing the Glucose Color Index variability metric value
}
\description{
Calculates Glucose Color Index (GCI)
}
\details{
Using weights, defaulting to those derived from analysis of the repeat
CGM wears in the HYPNOS trial, generates the GCI as a weighted linear combination
of the 6 degree of freedom disjoint piece-wise linear approximation of the
log-periodogram.
}
