% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pgram.R
\name{default_pgram}
\alias{default_pgram}
\title{Return reasonable defaults for CGM log-periodogram calculation}
\usage{
default_pgram()
}
\value{
options, list containing reasonable default options for pgram(), allowing
for user customization from baseline
}
\description{
Return reasonable defaults for CGM log-periodogram calculation
}
\details{
Defaults used in paper deriving the GCI (citation)
}
