% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCI.R
\name{default_weights}
\alias{default_weights}
\title{Return default weights for 6DF representation linear combination used to
calculte GCI}
\usage{
default_weights()
}
\value{
weights, vector containing the weights derived from the HYPNOS repeat
CGM wears
}
\description{
Return default weights for 6DF representation linear combination used to
calculte GCI
}
\details{
Weights derived using HYPNOS data, derived in paper introducing the
GCI (citation)
}
