% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{linear_approx}
\alias{linear_approx}
\title{Calculate 6 degree of freedom representation of a log-periodogram}
\usage{
linear_approx(pgram_obj)
}
\arguments{
\item{pgram_obj, }{object containing a log-transformed periodogram and its
frequency arguments}
}
\value{
pgram_summary, named vector containing the desired 6 degree of freedom
representation
}
\description{
Calculate 6 degree of freedom representation of a log-periodogram
}
\details{
Fits disjoint piece-wise linear models over 3 periodicity bands of
interest (greater than a day, a day to 150 minutes, shorter than 150 minutes),
returns the slopes and midpoints to form a 6-df representation of the full
periodogram
}
