% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{new_6DF}
\alias{new_6DF}
\title{Class for representing 6DF piece-wise linear model periodogram summary}
\usage{
new_6DF(long_params, int_params, short_params)
}
\arguments{
\item{long_params, }{named vector containing slope and midpoint over the
long-periodicity band of periodogram arguments}

\item{int_params, }{named vector containing slope and midpoint over the
intermediate-periodicity band of periodogram arguments}

\item{short_params, }{named vector containing slope and midpoint over the
short-periodicity band of periodogram arguments}
}
\value{
object with appropriate "6DF_summary" class and each summary
measure contained in its own component
}
\description{
Class for representing 6DF piece-wise linear model periodogram summary
}
\details{
Contains all slopes and midpoints across 3 periodicity bands of
interest (greater than a day, a day to 150 minutes, shorter than 150 minutes)
}
