% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{new_pgram}
\alias{new_pgram}
\title{Class for packaging a log-periodogram with its corresponding frequencies}
\usage{
new_pgram(log_periodogram, frequencies)
}
\arguments{
\item{log_periodogram, }{vector containing log-transformed periodogram ordinates}

\item{frequencies, }{vector containing frequency arguments corresponding to the
transformed periodogram values in "log_periodogram"}
}
\value{
object with appropriate "Periodogram" class and components containing
both the transformed periodogram and its frequency arguments
}
\description{
Class for packaging a log-periodogram with its corresponding frequencies
}
\details{
Object contains both the log-periodogram and frequencies as
corresponding vectors
}
