% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pgram.R
\name{pad_for_pgram}
\alias{pad_for_pgram}
\title{Pad the CGM time series with zeros}
\usage{
pad_for_pgram(ts, arg, options, consts, interval)
}
\arguments{
\item{ts, }{numeric time series of CGM data as a numeric vector}

\item{arg, }{POSIXct vector of timestamps for CGM time series}

\item{options, }{list containing relevant settings for periodogram calculation,
including number of expected days in the data ("days"), maximum gap in the
data allowable ("maxgap", in days), minimum data for analysis ("mindata" in days),
amount of start and end data to trim ("strim" and "etrim" both in days),
and odd-value spans of Daniel smoothing kernels ("spans")}

\item{consts, }{list of observation frequency constants}

\item{interval, }{time (in minutes) between CGM observations}
}
\value{
spec_data, vector containing the original CGM time series with
appropriate zero-padding (assuming sufficient data and continuity)
}
\description{
Pad the CGM time series with zeros
}
\details{
Ensures that glucose observations are appropriately and evenly
spaced, accounting for missing data within and potentially at the end of the
time series. Required step for accurate estimates of the spectral density
from the periodogram.
}
