% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pgram.R
\name{pgram}
\alias{pgram}
\title{Calculate CGM periodogram for a given time series and argument}
\usage{
pgram(ts, arg, options = default_pgram())
}
\arguments{
\item{ts, }{numeric time series of CGM data as a numeric vector}

\item{arg, }{POSIXct vector of timestamps for CGM time series}

\item{options, }{list containing relevant settings for periodogram calculation,
including number of expected days in the data ("days"), maximum gap in the
data allowable ("maxgap", in days), minimum data for analysis ("mindata" in days),
amount of start and end data to trim ("strim" and "etrim" both in days),
and odd-value spans of Daniel smoothing kernels ("spans"). This argument is
required, and a set of reasonable defaults is provided by the default_pgram()
function.}
}
\value{
"Periodogram" object with entries "lpgram" - vector of log-periodogram
values, "freqs" - vector of frequencies (min ^-1) corresponding to those values
}
\description{
Calculate CGM periodogram for a given time series and argument
}
\details{
This function calculates the CGM log-periodogram using appropriate
settings for this type of data
}
