% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{slope_midpoint}
\alias{slope_midpoint}
\title{Fit disjoint linear model portion and return slope/midpoint components}
\usage{
slope_midpoint(subset_df, freq_mid)
}
\arguments{
\item{subset_df, }{dataframe with subset of periodogram data corresponding to
a particular frequency band}

\item{freq_mid, }{frequency argument at which to evaluate the model to estimate
the periodogram midpoint over the given frequency band}
}
\value{
param_list, named list of summary parameters for the linear model over
the frequency band of interest
}
\description{
Fit disjoint linear model portion and return slope/midpoint components
}
\details{
Helper function to fit a linear model then extract the slope and
midpoint parameters
}
